improve(cli): move the remaining newCmd functions
